+++ src/main/java/br/com/softhouse/dende/model/dto/IngressoDTO.java
@@ -1,13 +1,14 @@
 package br.com.softhouse.dende.model.dto;
 
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record IngressoDTO(
         Long id,
         String nomeEvento,
         LocalDateTime dataEvento,
-        double valorPago,
+        BigDecimal valorPago,
         String status,
         LocalDateTime dataCompra
 ) {
